requirements-migrations.txt
alembic==1.14.0
asyncpg==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
